animations added & dark mode design implemented
