+++ README.md
@@ -1,3 +1,9 @@
-install d3 library
+install libraries
+
+d3 library
 
 npm install d3 --save
+
+framer-motion lbrary
+
+npm install framer-motion
