--- src/component/CustomizedMenus.css
@@ -1,4 +1,0 @@
-#dropdown-basic-button {
-   width: 160px;
-
-}
