+++ src/component/diseaseData/DataView.css
@@ -1,14 +1,11 @@
 .data-section {
-    background-color: rgb(173, 144, 201);
-    animation: slide 2s forwards;
+    background: rgb(37,35,36);
+    background: rgb(255,255,255);
+    background: linear-gradient(0deg, rgba(255,255,255,1) 8%, rgba(82,81,81,1) 20%, rgba(0,0,0,1) 45%);   animation: slide 2s forwards;
     animation-delay: 0s;
     height: 40rem;
 
 }
-
-/* .data-section[checking=true] {
-    height:35rem;
-} */
 
 @keyframes slide {
     100% { 
@@ -16,3 +13,6 @@
       }
 } 
 
+p {
+    color: white;
+}
