+++ src/component/diseaseData/DiseaseFilter.js
@@ -1,6 +1,6 @@
 import React, { useContext } from "react";
 import "./DiseaseFilter.css";
-import { AppContext } from "../../App";
+import { AppContext } from "../pages/DiseaseApp";
 
 export default function DiseaseFilter() {
   const {} = useContext(AppContext);
