+++ src/component/header/TitleHeader.css
@@ -4,11 +4,47 @@
 
 #header-container {
     display: flex;
-    background-color: aqua;
-    border: 2px solid black;
+    background: rgb(0,0,0);
+    /* background: linear-gradient(0deg, rgba(255,255,255,1) 0%, rgba(37,34,34,1) 5%);    border: 2px solid black; */
     justify-content: space-around;
+    align-items: center;
 }
 
-img {
+#header-container h1 {
+    color: white;
+    border-radius: 20%;
+    padding: .5em;
+    border: 2px white solid;
+}
+
+a {
+    text-decoration: none;
+}
+#home {
+    color: white;
+    font-size: 1.5em;
+    font-weight: bold;
+}
+
+#header-container img {
     width: 50px;
 }
+
+.logo:hover{
+    cursor: pointer;
+}
+
+#nav-items{
+    display: flex;
+    flex-direction: row;
+    justify-content: space-around;
+    gap: 1em;
+    align-items: center;
+}
+
+#diseases {
+    padding: .7rem;
+    border-radius: 50px;
+    background: rgb(255,255,255);
+    background: linear-gradient(94deg, rgba(255,255,255,1) 0%, rgba(160,172,255,1) 32%, rgba(55,79,255,1) 97%);}
+
