+++ src/component/map/ColorLegend.css
@@ -3,5 +3,9 @@
     color: black;
     margin-top: 25% ;
     margin-left: 10%;
-    border: 2px solid red;
+    border: 2px solid rgb(55, 79, 255);
+}
+
+#color-legend-container  p {
+color: black;
 }
