+++ src/component/map/Map.css
@@ -1,3 +1,7 @@
 #map-svg{
     flex: 1 1 auto;
 }
+/* #state-space{
+    background: rgb(255,255,255);
+background: linear-gradient(94deg, rgba(255,255,255,1) 0%, rgba(160,172,255,1) 32%, rgba(55,79,255,1) 97%);
+} */
