--- src/component/map/UsMap.css
@@ -1,4 +1,0 @@
-.svg-container{
-    width: 100%;
-    height: 100%;
-}
