+++ src/component/pages/DiseaseApp.css
@@ -1,0 +1,7 @@
+.main-container{
+    display: flex;
+    background: rgb(255,255,255);
+    background: linear-gradient(0deg, rgba(255,255,255,1) 0%, rgba(0,0,0,1) 100%);
+}
+
+
