+++ src/component/pages/Home.css
@@ -1,0 +1,3 @@
+h1 {
+color: black;
+}
